pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neuropilot"
version = "1.0.0"
description = "脑机接口康复系统（重构版）"
requires-python = ">=3.10"
dependencies = [
    "PyQt5>=5.15",
    "PyQt-Fluent-Widgets>=1.5",
    "pyqtgraph>=0.13",
    "matplotlib>=3.7",
    "numpy>=1.24",
    "scipy>=1.10",
    "scikit-learn>=1.3",
    "pandas>=2.0",
    "pyserial>=3.5",
    "SQLAlchemy>=2.0",
    "pydantic-settings>=2.0",
    "passlib[bcrypt]>=1.7",
    "bcrypt<4.1",
    "loguru>=0.7",
    "joblib>=1.3",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
lsl = ["pylsl>=1.16"]
bluetooth = ["pybluez2"]
dev = [
    "pytest>=7.0",
    "pytest-qt>=4.0",
    "ruff>=0.4",
    "mypy>=1.8",
    "pre-commit>=3.0",
    "types-passlib",
]

[project.scripts]
neuropilot = "neuropilot.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
files = ["src/neuropilot/domain", "src/neuropilot/infra"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
markers = [
    "qt: Qt UI tests",
]
